+++ apps/backoffice-v2/src/common/components/atoms/Select/Select.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import * as SelectPrimitive from '@radix-ui/react-select';
 
 export const Select = SelectPrimitive.Root;