+++ services/workflows-service/src/common/filters/AllExceptions.filter.ts
@@ -1,7 +1,7 @@
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 import { Catch, ArgumentsHost } from '@nestjs/common';
 import { BaseExceptionFilter, HttpAdapterHost } from '@nestjs/core';
-import { Response } from 'express';
+import type { Response } from 'express';
 
 @Catch()
 export class AllExceptionsFilter extends BaseExceptionFilter {
@@ -12,9 +12,11 @@ export class AllExceptionsFilter extends BaseExceptionFilter {
   catch(exception: unknown, host: ArgumentsHost) {
     // if (host.getType() === 'http') return;
     this.logger.error('Global error handler: ', exception as object);
-    super.catch(exception, host);
 
-    const response = host.switchToHttp().getResponse<Response>();
+    const context = host.switchToHttp();
+    const response = context.getResponse<Response>();
+
+    super.catch(exception, host);
 
     this.logger.error(`Outgoing response (Failure)`, {
       response: {