+++ services/workflows-service/src/events/workflow-state-changed-webhook-caller.ts
@@ -13,7 +13,7 @@ import { getWebhooks, Webhook } from '@/events/get-webhooks';
 import { sign } from '@/common/utils/sign/sign';
 import { env } from '@/env';
 import { CustomerService } from '@/customer/customer.service';
-import { TAuthenticationConfiguration } from '@/customer/types';
+import type { TAuthenticationConfiguration } from '@/customer/types';
 
 @Injectable()
 export class WorkflowStateChangedWebhookCaller {
@@ -52,11 +52,15 @@ export class WorkflowStateChangedWebhookCaller {
       'workflow.state.changed',
     );
 
-    const customer = await this.customerService.getByProjectId(data.runtimeData.projectId, {
-      select: {
-        authenticationConfiguration: true,
+    const customer = await this.customerService.getByProjectId(
+      // @ts-expect-error - error from Prisma types fix
+      data.runtimeData.projectId,
+      {
+        select: {
+          authenticationConfiguration: true,
+        },
       },
-    });
+    );
 
     const { webhookSharedSecret } =
       customer.authenticationConfiguration as TAuthenticationConfiguration;