+++ services/workflows-service/src/metrics/dto/get-users-assigned-cases-statistic.dto.ts
@@ -3,5 +3,6 @@ import { IsOptional } from 'class-validator';
 
 export class GetUsersAssignedCasesStatisticDto extends DateQueryParamsDto {
   @IsOptional()
+  // @ts-ignore
   fromDate!: Date;
 }